Struct devx_cmd::ChildProcess

pub struct ChildProcess { /* fields omitted */ }

Wraps std::process::Child, kills and waits for the process on Drop. It will log the fact that std::process::Child::kill was called in Drop. You should use wait for the process to finish with any of the available methods if you want to handle the error, otherwise it will be ignored.

Beware that ChildProcess holds an invariant that is not propagated to the type system. The invariant is that if ChildProcess was not spawned via Cmd::spawn_piped, then any methods that read the child's stdout will panic.

Implementations

impl ChildProcess

pub fn wait(&mut self) -> Result<()>

Waits for the process to finish. Returns an error if the process has finished with non-zero exit code.

You should use this method for processes spawned via Cmd::spawn since the output of the command won't be read and returned, but just written to this process's stdout (as stdout is inherited with Cmd::spawn)

pub fn read_bytes(self) -> Result<Vec<u8>>

Same as ChildProcess::read but reads any bytes sequence from the child process stdout.

Panics

Same as for ChildProcess::read.

pub fn read(self) -> Result<String>

Waits for the process to finish and returns all that it has written to stdout. Returns an error if the process has finished with non-zero exit code. Expects a valid utf8 bytes sequence (since it returns a Rust String), if the process is not guaranteed to output valid utf8 you might want to use ChildProcess::read_bytes instead.

If Cmd::echo_cmd has been set to true then prints captured output to stderr.

Panics

Panics if the process was spawned with non-piped stdout. This method is expected to be used only for processes spawned via Cmd::spawn_piped.

pub fn stdout_lines(&mut self) -> impl Iterator<Item = String> + '_

Returns an iterator over the lines of data output to stdout by the child process. Beware that the iterator buffers the output, thus when the it is dropped the buffered data will be discarded and following reads won't restore it.

Panics

Panics if some std::io::Error happens during the reading. All invariants from [ChildProcess::read_bytes] apply here too.